Bootstrap resample an information measure

Runs a simple bootstrap resampling of your observations on information measure of choice to assess uncertainty.

bootstrap_info_interval(
  a_vec,
  b_vec,
  measure_func = entropynet::calc_mutual_info,
  B = 500,
  ...
)

Arguments

a_vec
an atomic vector of observations (can be logical, factor, character, or integer).

b_vec
paired observations for second variable.

measure_func
Any function that takes as its first two arguments to vectors of categorical values and returns a number can be used as a "measure".

B
Number of bootstrap resamples to gather

...
Any additional arguments to measure_func outside of a_vec and b_vec.

Value

An array of values returned from running measure_func on bootstrap resampled observations from provided data.

Examples

N <- 100
pois_mi_samples <- bootstrap_info_interval(
  a_vec = rpois(N, 5),
  b_vec = rpois(N, 8),
  B = 150
)

#> Error in sample(1:N, replace = TRUE): object 'N' not found

hist(pois_mi_samples)

#> Error in hist(pois_mi_samples): object 'pois_mi_samples' not found

# Any function that takes as its first two arguments to vectors of categorical
# values and returns a number can be used as a "measure".
pois_je_samples <- bootstrap_info_interval(
  a_vec = rpois(N, 5),
  b_vec = rpois(N, 8),
  measure = calc_joint_ent,
  B = 150
)

#> Error in sample(1:N, replace = TRUE): object 'N' not found

hist(pois_je_samples)

#> Error in hist(pois_je_samples): object 'pois_je_samples' not found

# You can pass any arguments to the entropy function you may desire as well...
binom_mi_samples <-
  bootstrap_info_interval(
    a_vec = rbinom(N, size = 1, p = 0.1),
    b_vec = rbinom(N, size = 1, p = 0.08),
    B = 150,
    measure = function(a, b, scalar) calc_joint_ent(a, b)*scalar,
    scalar = -2
  )

#> Error in sample(1:N, replace = TRUE): object 'N' not found

hist(binom_mi_samples)

#> Error in hist(binom_mi_samples): object 'binom_mi_samples' not found

#'