Build counts observation vec

Build counts observation vec

build_count_vec(n, fill_indices, fill_values)

Arguments

n
Total number of observations (aka the length of the final vector).

fill_indices
Integer vector of all indices an occurrence happened.

fill_values
Array of same length as fill_indices that contain the integer counts for the filled observations.

Value

Array of length n with every observation indicated by fill_indices filled with the matching entry in fill_values and 0 elsewhere.

Examples

build_count_vec(10, c(1,3,4), fill_values = c(10, 20, 30))

#>  [1] 10  0 20 30  0  0  0  0  0  0