Calculate entropy network

Takes a dataset of occurence or count observations and runs a pair-wise information network over all combinations of variables.

calc_entropy_net(
  pairs,
  id_col,
  target_col,
  count_col,
  info_func = entropynet::calc_mutual_info,
  possible_targets,
  parallel = TRUE,
  subset_pairs = FALSE,
  verbose = FALSE
)

Arguments

pairs
Dataframe with at least two columns: id and target. id corresponds to the variables to be associated with each other and target corresponds to the observations they had. This is in essence a tidy version of an tabular dataset with columns as id and rows as target. If data is count based an addition count column can be included. (Not yet implemented).

id_col
Name of column containing variable id (I.e. nodes of the association network).

target_col
Name of the column containing the target for occurance (I.e. the data used to infer association between nodes).

count_col
Name of column containing count values as integers. If this is not set and no column with title count is included in pairs then entropy will be run on binary occurrence.

info_func
What information function should be used? Like all of the calc_* functions in this package this takes as its arguments two paired vectors of observations as input and returns a number corresponding to strength of association.

possible_targets
If there are some target values that are never seen in your dataset then they can be provided here with the form of a characterr, factor, or integer vector of all target values.

parallel
Should the processing be run in parallel? Three options. A logical TRUE will run in parallel over every core available using future::multisession. A number will run in parallel over that many cores. E.g. parallel = 5 will run in parallel on 5 core. Last, for the most control a future processing plan can be passed (I.e. what is run by future::plan() to set up your parallel processing environment.) This allows you to run on clusters etc..

subset_pairs
Should the algorithm be run on a random subset of possible combinations? Useful for making sure everything is good before running larger jobs. Set to the number of pairs you want to calculate strength for. If reproducability is desired use set.seed().

verbose
Information about steps is written to console.

Value

Examples

# Run association net on the virus-host dataset
calc_entropy_net(
  pairs = virus_host,
  id_col = virus_id,
  target_col = host_id,
  parallel = FALSE,
  subset_pairs = 100
)

#> # A tibble: 100 x 3
#>    a                                    b                               strength
#>    <chr>                                <chr>                              <dbl>
#>  1 Human papillomavirus 60              Manzanilla virus              0.00000254
#>  2 Caprine arthritis encephalitis virus Manzanilla virus              0.00000509
#>  3 Bovine ephemeral fever virus         Wesselsbron virus             0.00313   
#>  4 Human papillomavirus 53              Swinepox virus                0.00000254
#>  5 Lake Victoria marburgvirus           Murine adenovirus A           0.0000179 
#>  6 Australian bat lyssavirus            Cercopithecine herpesvirus 16 0.000116  
#>  7 Hippotragine herpesvirus 1           Rubella virus                 0.00000254
#>  8 California encephalitis virus        Human parainfluenza virus 1   0.00421   
#>  9 Tamana bat virus                     Equine adenovirus A           0.0000333 
#> 10 Japanese encephalitis virus          Moloney murine sarcoma virus  0.0000672 
#> # … with 90 more rows

# Run count-based association for vir-clasp rna-seq dataset
calc_entropy_net(
  pairs = virclasp_chikv,
  id_col = id,
  target_col = condition,
  count_col = count,
  parallel = FALSE,
  subset_pairs = 100
)

#> # A tibble: 100 x 3
#>    a           b           strength
#>    <chr>       <chr>          <dbl>
#>  1 RFT1_HUMAN  MBB1A_HUMAN   1.94  
#>  2 SAHH2_HUMAN SAE2_HUMAN    0.969 
#>  3 PSB5_HUMAN  TGM3_HUMAN    0.0766
#>  4 OST48_HUMAN OTUD4_HUMAN   0.503 
#>  5 GLYG_HUMAN  KPRP_HUMAN    1.15  
#>  6 NUP93_HUMAN APBB2_HUMAN   0.0786
#>  7 ARHGB_HUMAN AUTS2_HUMAN   0.0839
#>  8 LTN1_HUMAN  CMTR1_HUMAN   2.74  
#>  9 SYMC_HUMAN  PLPL6_HUMAN   1.70  
#> 10 ABCB7_HUMAN PSMF1_HUMAN   0.595 
#> # … with 90 more rows